Separate domain logic from plugin class
- this is needed as the dependency on obsidian makes testing impossible, this
should facilitate me stubbing the methods I want and testing everything else
