+++ src/obsidianHelpers.js
@@ -1,0 +1,4 @@
+import { Notice } from 'obsidian';
+import { getDailyNoteSettings, getAllDailyNotes, getDailyNote } from 'obsidian-daily-notes-interface'
+
+export { Notice, getDailyNoteSettings, getAllDailyNotes, getDailyNote }
